requirements.txt
streamlit
requests
pandas
faster-whisper
pyspellchecker
python-docx